Fix the shake drop for fp16 training

+++ researchlib/layers/shakedrop.py
@@ -49,23 +49,23 @@ class ShakeDropFunction(torch.autograd.Function):
             raise ValueError("beta_range should be two-element list or 0")
 
         if training:
-            gate = torch.cuda.FloatTensor([0]).bernoulli_(p)
+            gate = torch.cuda.FloatTensor([0]).to(x.dtype).bernoulli_(p)
             beta_range = torch.tensor(beta_range)
             ctx.save_for_backward(mode, gate, beta_range)
             if gate.item() == 0:
                 if type(alpha_range) == list:  # two-element list
                     if int(mode) == 0:
                         alpha = torch.FloatTensor(x.size(0)).to(
-                            x.device).uniform_(*alpha_range)
+                            x.device).to(x.dtype).uniform_(*alpha_range)
                         alpha = alpha.view(alpha.size(0), 1, 1, 1).expand_as(x)
                     elif int(mode) == 1:
                         alpha = torch.FloatTensor(x.size(0), x.size(1)).to(
-                            x.device).uniform_(*alpha_range)
+                            x.device).to(x.dtype).uniform_(*alpha_range)
                         alpha = alpha.view(alpha.size(0), alpha.size(1), 1,
                                            1).expand_as(x)
                     elif int(mode) == 2:
                         alpha = torch.empty_like(x).to(
-                            x.device).uniform_(*alpha_range)
+                            x.device).to(x.dtype).uniform_(*alpha_range)
                 elif alpha_range == 0:
                     alpha = 0
                 return alpha * x
@@ -84,18 +84,18 @@ class ShakeDropFunction(torch.autograd.Function):
             if len(beta_range) == 2:  # two-element list
                 if int(mode) == 0:
                     beta = torch.FloatTensor(grad_output.size(0)).to(
-                        grad_output.device).uniform_(*beta_range)
+                        grad_output.device).to(grad_output.dtype).uniform_(*beta_range)
                     beta = beta.view(beta.size(0), 1, 1,
                                      1).expand_as(grad_output)
                 elif int(mode) == 1:
                     beta = torch.FloatTensor(
                         grad_output.size(0), grad_output.size(1)).to(
-                            grad_output.device).uniform_(*beta_range)
+                            grad_output.device).to(grad_output.dtype).uniform_(*beta_range)
                     beta = beta.view(beta.size(0), beta.size(1), 1,
                                      1).expand_as(grad_output)
                 elif int(mode) == 2:
                     beta = torch.empty_like(grad_output).to(
-                        grad_output.device).uniform_(*beta_range)
+                        grad_output.device).to(grad_output.dtype).uniform_(*beta_range)
                 beta = Variable(beta)
             elif beta_range == 0:
                 beta = 0
